fix: corrigir erros do linter
Backend:
- Remover imports não utilizados (Request, DbUser, requireAdmin, UserUpdateInput)
- Corrigir variável password não utilizada em authService.ts
- Utilizar validação updateUserValidation que estava definida mas não usada
- Remover authService não utilizado em permissions.ts

Frontend:
- Corrigir aspas não escapadas em AccessDenied.tsx (&quot;Roles&quot;)
- Separar hook useAuth em arquivo próprio para resolver warning do Fast Refresh
- Atualizar imports do useAuth em todos os componentes

Resultado: 0 erros críticos em backend e frontend
