+++ frontend/src/components/Header.tsx
@@ -1,4 +1,4 @@
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../hooks/useAuth';
 import { LogOut, User } from 'lucide-react';
 
 export const Header: React.FC = () => {
